fix table render
